+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changes
 
 - Color outputs.
+- Ask for confirmation when using `delete` subcommand.
 
 ### Fixed